pyproject.toml
[tool.poetry]
name = "flask-mongo"
version = "0.1.0"
description = "A flask app for taking better notes while learning shas (Talmud)."
authors = ["DewofyourYouth <jacobshore@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^1.1.2"
pymongo = "^3.11.0"
python_dotenv = "^0.14.0"
flask-wtf = "^0.14.3"
flask-login = "^0.5.0"
Flask-PyMongo = "^2.3.0"
bcrypt = "^3.2.0"
flask_pymongo = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
rope = "^0.18.0"
black = "^20.8b1"
bpython = "^0.20"
pytest-mongodb = "^2.2.0"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
